move to neural module
