+++ src/moscot/neural/problems/__init__.py
@@ -1,0 +1,3 @@
+from moscot.neural.problems.generic import GENOTLinProblem
+
+__all__ = ["GENOTLinProblem"]
