+++ src/moscot/neural/problems/generic/__init__.py
@@ -1,0 +1,3 @@
+from moscot.neural.problems.generic._generic import GENOTLinProblem
+
+__all__ = ["GENOTLinProblem"]
