+++ src/moscot/problems/__init__.py
@@ -1,5 +1,4 @@
 from moscot.problems.cross_modality import TranslationProblem
-from moscot.problems.generic import GENOTLinProblem
 from moscot.problems.space import AlignmentProblem, MappingProblem
 from moscot.problems.spatiotemporal import SpatioTemporalProblem
 from moscot.problems.time import LineageProblem, TemporalProblem
@@ -11,5 +10,4 @@ __all__ = [
     "SpatioTemporalProblem",
     "LineageProblem",
     "TemporalProblem",
-    "GENOTLinProblem",
 ]
