+++ src/moscot/problems/generic/__init__.py
@@ -1,14 +1,8 @@
-from moscot.problems.generic._generic import (
-    FGWProblem,
-    GENOTLinProblem,
-    GWProblem,
-    SinkhornProblem,
-)
+from moscot.problems.generic._generic import FGWProblem, GWProblem, SinkhornProblem
 from moscot.problems.generic._mixins import GenericAnalysisMixin
 
 __all__ = [
     "FGWProblem" "SinkhornProblem",
-    "GENOTLinProblem",
     "GWProblem",
     "GenericAnalysisMixin",
 ]
